Initial commit.
Unless otherwise specified, the contents of this project are
provided under the following licenses:

The contents of boards/* is provided under the CC-BY-3.0 license.

    https://creativecommons.org/licenses/by/3.0/

This includes the graphical assets, JSON declarations, markdown
documentation files, etc.

The Board Explorer server and front-end, including the Javascript, HTML,
and CSS, unless otherwise noted in the source files, are provided
under the Apache 2.0 license:
---
Copyright (C) 2017 Intel Corporation

Licensed under the Apache License, Version 2.0 (the "License");
you may not use this file except in compliance with the License.
You may obtain a copy of the License at

    http://www.apache.org/licenses/LICENSE-2.0

Unless required by applicable law or agreed to in writing, software
distributed under the License is distributed on an "AS IS" BASIS,
WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
See the License for the specific language governing permissions and
limitations under the License.
---

The routines used to parse and inject the SVG into the DOM are from
the [SVGInjector project](https://github.com/iconic/SVGInjector) and
are Copyright (c) 2014-2015 Waybury, licensed under the MIT license.

Signed-off-by: James Ketrenos <james.p.ketrenos@intel.com>